refactor: 반환 타입 long → int 통일 및 입장 로직에 active 유저 좌석 차감 적용

+++ src/main/java/kr/jemi/zticket/seat/adapter/in/web/dto/AvailableCountResponse.java
@@ -1,4 +1,4 @@
 package kr.jemi.zticket.seat.adapter.in.web.dto;
 
-public record AvailableCountResponse(long count) {
+public record AvailableCountResponse(int count) {
 }
